Small function refactor.

+++ GasolinaDataUpdater/Json.cs
@@ -16,14 +16,16 @@ namespace DownloaderLibrary
         {
             try
             {
-
-                List<PriceDTO> precios = ExcelFileReader.Read(excelFile);
                 string fileName = Path.GetFileNameWithoutExtension(excelFile) + "estados.json";
                 string fileDestination = Path.Combine(jsonDirectory, fileName.Substring(4, 2), fileName);
+
+                if (File.Exists(fileDestination))
+                    return string.Empty;
 
                 if (Directory.Exists(Path.GetDirectoryName(fileDestination)) == false)
                     Directory.CreateDirectory(Path.GetDirectoryName(fileDestination));
 
+                List<PriceDTO> precios = ExcelFileReader.Read(excelFile);
                 string estadosJson = GenerateJsonEstados(precios, Path.GetFileNameWithoutExtension(excelFile).Substring(0, 8));
                 File.WriteAllText(fileDestination, estadosJson);
